Added FTP REST (resume transfer) capability

+++ ftp.h
@@ -1,6 +1,6 @@
 #ifndef FTP_H
 #define FTP_H
 
-int ftp_RETR(int sock, const char *file, FILE **out, long fpos);
+int ftp_RETR(int sock, const char *file, FILE **out, size_t fpos);
 
 #endif
